+++ artemis-weaver/src/main/java/com/artemis/meta/FieldDescriptor.java
@@ -1,5 +1,7 @@
 package com.artemis.meta;
 
+import org.objectweb.asm.tree.AbstractInsnNode;
+
 public class FieldDescriptor {
 	public int access;
 	public String name;
@@ -7,6 +9,7 @@ public class FieldDescriptor {
 	public String signature;
 	public Object value;
 	public int offset; // byte offset, only used by @PackedComponent
+	public AbstractInsnNode reset;
 
 	public FieldDescriptor(int access, String name, String desc, String signature, Object value) {
 		this.name = name;
@@ -23,5 +26,18 @@ public class FieldDescriptor {
 
 	public FieldDescriptor(String name) {
 		this.name = name;
+	}
+
+	public boolean isResettable() {
+		return desc.length() == 1 || "Ljava/lang/String;".equals(desc);
+	}
+
+	@Override
+	public String toString() {
+		return "FieldDescriptor{" +
+			"name='" + name + '\'' +
+			", desc='" + desc + '\'' +
+			", signature='" + signature + '\'' +
+			'}';
 	}
 }