+++ artemis-weaver/src/test/java/com/artemis/ComponentTypeWeaverTest.java
@@ -1,7 +1,6 @@
 package com.artemis;
 
 import static com.artemis.Transformer.transform;
-import static com.artemis.Weaver.scan;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -9,7 +8,6 @@ import static org.junit.Assert.assertTrue;
 import java.io.InputStream;
 
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.Type;
@@ -18,7 +16,6 @@ import com.artemis.component.ComponentToWeave;
 import com.artemis.component.IllegalComponent;
 import com.artemis.component.PackedToBeB;
 import com.artemis.component.PooledForced;
-import com.artemis.component.PooledNotForced;
 import com.artemis.component.SingletonTagComponent;
 import com.artemis.meta.ClassMetadata;
 import com.artemis.meta.ClassMetadata.GlobalConfiguration;