+++ artemis-weaver/src/test/java/com/artemis/component/DefaultBoolean.java
@@ -1,6 +1,0 @@
-package com.artemis.component;
-
-public class DefaultBoolean {
-	public boolean set = true;
-	public boolean unset;
-}