+++ artemis-weaver/src/test/java/com/artemis/component/DefaultDecimal.java
@@ -1,8 +1,0 @@
-package com.artemis.component;
-
-public class DefaultDecimal {
-	public float f0 = 0;
-	public double d0 = 0;
-	public float f123 = 123f;
-	public double d123 = 0.00123f;
-}