+++ artemis-weaver/src/test/java/com/artemis/component/DefaultInteger.java
@@ -1,9 +1,0 @@
-package com.artemis.component;
-
-public class DefaultInteger {
-	public byte b012 = 0x12;
-	public byte b0 = 0;
-	public int i0 = 0;
-	public int i543 = 543;
-	public long l1 = -1;
-}